+++ src/__tests__/PictureList.tsx
@@ -1,14 +1,15 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import App from '../../App';
-import store from '../../redux/store';
 import { Provider } from 'react-redux';
+
+import PictureList from '../components/PictureList';
+import store from '../redux/store';
 
 describe('<SinglePicture>', () => {
   it('Should render component', () => {
     const wrapper = shallow(
       <Provider store={store}>
-        <App />
+        <PictureList />
       </Provider>
     );
     expect(wrapper).toMatchSnapshot();
